MINIFICPP-1606 ProcessSession::read() should return int64_t
Since ProcessSession::read() returns the size of the flow file, and its return
type was int, we still couldn't handle flow files larger than around 2 GB.

+++ libminifi/include/serialization/FlowFileSerializer.h
@@ -42,11 +42,11 @@ class InputStreamCallback;
 
 class FlowFileSerializer {
  public:
-  using FlowFileReader = std::function<int(const std::shared_ptr<core::FlowFile>&, InputStreamCallback*)>;
+  using FlowFileReader = std::function<int64_t(const std::shared_ptr<core::FlowFile>&, InputStreamCallback*)>;
 
   explicit FlowFileSerializer(FlowFileReader reader) : reader_(std::move(reader)) {}
 
-  virtual int serialize(const std::shared_ptr<core::FlowFile>& flowFile, const std::shared_ptr<io::OutputStream>& out) = 0;
+  virtual int64_t serialize(const std::shared_ptr<core::FlowFile>& flowFile, const std::shared_ptr<io::OutputStream>& out) = 0;
 
   virtual ~FlowFileSerializer() = default;
